refactor(snowflake): keep SPCS auth in cloud package
